+++ linodecli/configuration/helpers.py
@@ -7,7 +7,6 @@ import os
 import webbrowser
 from typing import Any, Callable, List, Optional
 
-from .auth import _do_get_request
 
 LEGACY_CONFIG_NAME = ".linode-cli"
 LEGACY_CONFIG_DIR = os.path.expanduser("~")
@@ -174,13 +173,25 @@ def _default_thing_input(
 
 def _default_text_input(
     ask: str,
-    default: str = None,
+    default: Optional[str] = None,
     optional: bool = False,
     validator: Callable[[str], Optional[str]] = None,
 ) -> Optional[str]:  # pylint: disable=too-many-arguments
     """
     Requests the user to enter a certain string of text with the given prompt.
     If optional, the user may hit enter to not configure this option.
+
+    :param ask: The initial question to ask the user.
+    :type ask: str
+    :param default: The default value for this input.
+    :type default: Optional[str]
+    :param optional: Whether this prompt is optional.
+    :type optional: bool
+    :param validator: A function to validate the user's input with.
+    :type validator: Callable[[str], Optional[str]]
+
+    :returns: The user's input.
+    :rtype: str
     """
 
     prompt_text = f"\n{ask} "
@@ -218,9 +229,17 @@ def _default_text_input(
 
 def _bool_input(
     prompt: str, default: bool = True
-):  # pylint: disable=too-many-arguments
+) -> bool:  # pylint: disable=too-many-arguments
     """
     Requests the user to enter either `y` or `n` given a prompt.
+
+    :param prompt: The prompt to ask the user.
+    :type prompt: str
+    :param default: The default value for this input. Defaults to True.
+    :type default: bool
+
+    :returns: The user's input.
+    :rtype: bool
     """
     while True:
         user_input = input(f"\n{prompt} [y/N]: ").strip().lower()
@@ -240,86 +259,20 @@ def _config_get_with_default(
     user: str,
     field: str,
     default: Any = None,
-) -> Optional[Any]:
+) -> Any:
     """
     Gets a ConfigParser value and returns a default value if the key isn't found.
+
+    :param user: The user to get a value for.
+    :type user: str
+    :param field: The name of the field to get the value for.
+    :type field: str
+    :param default: The default value to use if a value isn't found. Defaults to None.
+    :type default: Any
+
+    :returns: The value pulled from the config or the default value.
+    :rtype: Any
     """
     return (
         config.get(user, field) if config.has_option(user, field) else default
     )
-
-
-def _handle_no_default_user(self):  # pylint: disable=too-many-branches
-    """
-    Handle the case that there is no default user in the config
-    """
-    users = [c for c in self.config.sections() if c != "DEFAULT"]
-
-    if len(users) == 1:
-        # only one user configured - they're the default
-        self.config.set("DEFAULT", "default-user", users[0])
-        self.write_config()
-        return
-
-    if len(users) == 0:
-        # config is new or _really_ old
-        token = self.config.get("DEFAULT", "token")
-
-        if token is not None:
-            # there's a token in the config - configure that user
-            u = _do_get_request(
-                self.base_url, "/profile", token=token, exit_on_error=False
-            )
-
-            if "errors" in u:
-                # this token was bad - reconfigure
-                self.configure()
-                return
-
-            # setup config for this user
-            username = u["username"]
-
-            self.config.set("DEFAULT", "default-user", username)
-            self.config.add_section(username)
-            self.config.set(username, "token", token)
-
-            config_keys = (
-                "region",
-                "type",
-                "image",
-                "mysql_engine",
-                "postgresql_engine",
-                "authorized_keys",
-                "api_host",
-                "api_version",
-                "api_scheme",
-            )
-
-            for key in config_keys:
-                if not self.config.has_option("DEFAULT", key):
-                    continue
-
-                self.config.set(username, key, self.config.get("DEFAULT", key))
-
-            self.write_config()
-        else:
-            # got nothin', reconfigure
-            self.configure()
-
-        # this should be handled
-        return
-
-    # more than one user - prompt for the default
-    print("Please choose the active user.  Configured users are:")
-    for u in users:
-        print(f" {u}")
-    print()
-
-    while True:
-        username = input("Active user: ")
-
-        if username in users:
-            self.config.set("DEFAULT", "default-user", username)
-            self.write_config()
-            return
-        print(f"No user {username}")
